Move files index writer to file_interface
