+++ test/end_to_end/test_create_run_from_mesh.py
@@ -68,48 +68,3 @@ def test_create_run_from_mesh_outcrop_explicit_files(tmp_path, end_to_end_fixtur
         'new_run.fehm',
         'new_run.files',
     }
-
-
-def test_create_files_index(tmp_path):
-    run_directory = tmp_path / 'run'
-    run_directory.mkdir()
-
-    files_config = FilesConfig(
-        run_root='run_root',
-        material_zone='material_zone.txt',
-        outside_zone='outside_zone.txt',
-        area='area.txt',
-        rock_properties='rock_properties.txt',
-        conductivity='conductivity.txt',
-        pore_pressure='pore_pressure.txt',
-        permeability='permeability.txt',
-        heat_flux='heat_flux.txt',
-        flow='flow.txt',
-        files='files.txt',
-        grid='grid.txt',
-        input='input.txt',
-        output='output.txt',
-        store='store.txt',
-        history='history.txt',
-        water_properties='water_properties.txt',
-        check='check.txt',
-        error='error.txt',
-        final_conditions='final_conditions.txt',
-    )
-    create_files_index(run_directory, files_config)
-    files_index_file = run_directory / 'files.txt'
-    assert files_index_file.read_text() == (
-        'root: run_root\n'
-        'input: input.txt\n'
-        'outpu: output.txt\n'
-        'grida: grid.txt\n'
-        'storo: store.txt\n'
-        'rsto: final_conditions.txt\n'
-        'error: error.txt\n'
-        'check: check.txt\n'
-        'zone: material_zone.txt\n'
-        'look: water_properties.txt\n'
-        'hist: history.txt\n\n'
-        'all'
-    )
-
